IAM: Workaround for hibernate issue that embedded map in an entity ignore key cases.
Looks like Hibernate has a problem persist an embedded map in an entity. It ignores the
map key case. So if you add both 'email' and 'Email' to the user info map, it causes a
"duplicate entry ... for key 'PRIMARY'" failure. This workaround makes all user info
keys case insensitive, which also makes sense, so that 'email' and 'Email' map to the
same info entry.