Fixes 1733 - Allows GET ACL of versioned object and fixes version delete
Allows authorized users to get ACL of versioned object even when versioning is suspended on the bucket. Previously this operation was not allowed when versioning is suspended.

Also fixes issue where anyone can delete a previous object version when versioning is suspended. According to S3 docs, only the bucket owner can delete specific object versions
when versioning is suspended on the bucket.