Add tokens module for AWS STS temporary creds.
An authentication pipeline for use with account name, username and
password is added to support temporary credentials for the UI proxy. A
QueryPipeline support class is now used for any services supporting
query access. There is a new Ciphers utility class for users of the JDK
Cipher API.

See amazon docs for more info:
http://docs.amazonwebservices.com/STS/latest/UsingSTS/Welcome.html