Reporting updates related to instance usage.
Sensor data retrieved from the Cluster Controller is now filtered to
remove all but the latest value for any metric prior to propagation.
Unnecessary fields have been removed from the DescribeSensors message.

The reporting instance usage event is updated for clarity (renamed the
resourceName field to instanceId) and to remove the unused timestamp.
The sequenceNum is changed to a Long to match the incoming message type.

InstanceUsageArtEntity fields now reflects the data being retrieved from
instances. The InstanceArtGenerator is updated to populate data for the
type usage summaries and to exclude instances with no data in the report
period. The InstanceRenderer is updated to display the additional fields
and to display '0' rather than '(null)' for instance type usage
summaries with no usage. A InstanceArtGeneratorTest has been added for
unit testing of the InstanceArtGenerator.

An AbstractArtGenerator is introduced for any common generation
functionality. This is used for Elastic IPs and Instances.

Instance creation and usage events are updated to remove redundant
fields.

The unused ReportingInstanceAttributeEvent class is removed.

Configuration settings for sensor value polling have been moved to the
DescribeSensorsListener class from the Units class.