Adds Walrus capacity configurable property value for reporting Walrus storage usage.
Walrus capacity is configured initially to be the space available in
$eucalyptus_home/var/lib/eucalyptus/bukkits minus 1GB (if result > 0).
If it has been configured before (i.e. a DB entry for the walrus install
exists) it will use that value. New configurable property is:
walrus.storagemaxtotalcapacity and can be changed at runtime using
euca-modify-properties.

NOTE: no enforcement of this value is done yet, this is purely for reporting
and therefore usage in reports may exceed 100%.