Make key-syncing code explicitly specify paths
Previously, key-syncing code took a list of directories and a list of
file names and searched the directories for files with names which
are in the given list, with crippling consequences in multi-cluster
deployments.  This fix makes every key-syncing call explicitly say
where to look for the files that it wants.

Fixes EUCA-2982, for real this time.