Persistence context connectivity exception handling updated.
The PersistenceContextLookupInterceptor is now generalized to allow
handling of other events. The PersistenceExceptions class is updated to
notify event interceptors on connectivity errors and the existing
DatabasesPersistenceContextLookupInterceptor now handles this condition.

This decoupling allows use of PersistenceContexts without the
dependencies that are necessary if the Databases class is used (such as
HA-JDBC)