Reporting data export format and streaming changes.
MessageStackHandler is updated to have a default empty implementation
for 'outgoingMessage'. Subclasses are updated to remove their own empty
implementations of this previously abstract method.

Binding is updated to support streaming using new 'toStream' /
'fromStream' methods. Areas that previously converted to an OM model and
then immediately to bytes are updated to skip the intermediate step. The
namespace processing that was previously performed using 'replaceAll' is
now performed via a new NamespaceMappingXMLStreamWriter class.

The reporting Export class is modified to support streaming exports. The
'eureport-export-data' command is updated to support compression and
streaming. A patched version of the Python gzip module is included to
work around a bug that prevents streaming of gzip data (fixed in Python
3.2). The reporting Import class is modified to support streaming import
of data. The RestfulMarshallingHandler has a new 'streamResponse' method
that can be called from service implementations to stream data. In this
case the method will return 'null' to indicate that the response has
been handled. The 'streamResponse' method uses HTTP chunked encoding and
supports gzip and deflate data compression.

A new 'eureport-delete-data' command has been added and can be used to
delete (presumably exported) reporting data. The Entities class has been
updated to support bulk deletion of data. The reporting messages,
binding and service are updated to support reporting data deletion.

An EventFactory class has been added for creation of persistent
reporting events from exported data. ExportUtils is updated to provide
functions mapping between the export model and the internal persistence
events. The new ReportedAction and ReportedUsage classes are the
external representation of the internal reporting persistence events. A
JiBX binding has been added for reporting exports and is imported for
use in the reporting service binding. A new ReportingEventIdGenerator
has been added to allow exported reporting data to be saved to the data
warehouse with the original identifiers (so data can be merged). The
ReportingEventSupport no longer extends AbstractPersistent, this means
that we no longer have various unused fields/columns and that we can
preserve the original identifier and created timestamp for each event.

EventRecord has been updated to allow reuse of eucalyptus code
in the data warehouse without depending on Mule classes.