Refactor of storage controller and storage san code
The refactor moved the DASManager and associated entity classes into the
Eucalyptus tree out of the proprietary code. There are now two available
storage controller backends: the OverlayManager and DASManager. The
DASManager provides Direct-Attached-Storage capabilities via LVM
manipulation of disks to provide EBS volumes and snapshots.

The StorageController now has a new lifecycle configuration and is
instantiated during a fireLoad() call from the bootstrapper which occurs
after the normal load() bootstrap cycle. This guarantees that the DB and
entity layers are available at the SC bootstrap time.

The BlockStorageManagerFactory is no longer used to build the backend
LogicalStorageManager, the StorageManagers class builds a list of classes
during compile time base on a new annotation for storage managers:
@StorageManagerProperty(name,manager). If the annotated class
implements the LogicalStorageManager interface it need only specify the
name (i.e. OverlayManager => @StorageManagerProperty("overlay")), but
classes that implement something other than LogicalStorageManager but are
providers which would be loaded by a LogicalStorageManager must include
both their identifier name and the name of the LogicalStorageManager that
loads them
(eg NetappProvider => @StorageManagerProperty("Netapp",manager="SANManager")

The backend for a SC is now defined at runtime by a modifiable property:
<Parition>.storage.blockstoragemanager

e.g. euca-modify-property -p PARTI00.storage.blockstoragemanager=das

The configuration only exists for the registration lifecycle of the SC
component, so to change the configuration you must deregister and
re-register the SC. This is intended to strongly discourage changing
of backends once one is configured. However, the SC does *NOT* default
to any backend and must be explicitly configured to use a backend after
being registered.