Reporting and general upgrade changes for 3.2 (EUCA-3828)
Reporting upgrade is updated to use the new upgrade mechanism. Three
new classes are added for handling reporting upgrade to 3.2. The
PreUpgrade32 class removes reporting tables that are no longer used.
The Upgrade32 class runs for reporting user and account entities and
removes the existing data (which will be missing some columns). The
PostUpgrade32 class populates reporting data in the new format for any
items that exist in the cloud. The UpgradeUtils class provides
supporting methods for reporting upgrades.

Upgrade scripts and support classes related to the old upgrade mechanism
are removed. The StandalonePersistence class is preserved as it can be
used when running test code (it is no longer required for upgrade)