Fixes SC configuration problem when SC is remote from CLC. Fixes EUCA-3987.
fireChange() listener for blockstoragemanager now looks up the configs for all SCs in the partition and checks for the validity of the new value in the union of all the valid backends. Adds a list of valid values to the StorageController service config such that the CLC can validate input on that list. List is populated at load time of the service.