Fix for cloud controller persistence of address allocations (EUCA-2427)
Add/remove of 'Address' entities is updated to use 'EntityTransaction'
rather than 'EntityWrapper'. This change allows use of a JPA merge when
persisting the entity, which avoids making the 'cache' copy of each
'Address' entity persistent.

When loading 'Address' information from the database a new method is
used ('resetPersistence') which clears the 'low level' persistent
information for an entity. Clearing this information means that when the
entity is next persisted it will have a new ID, version and
created/updated dates.

When adding an 'Address' we clear the natural ID so a new one is created
for the allocation.

Fixes EUCA-2427