Extend evil roboto-correcting hack to deregister
euca-deregister-* commands for non-NC components require one to supply
names for the new components as positional args.  When they aren't
given, roboto's error message recommends using the --name option,
which doesn't exist.  This patch removes the hyphens from the front
of that option to make the error output more reasonable.

This is the same hack as the one introduced in commit e208a5ef,
but applied to de-registration commands instead.

Fixes EUCA-1109