Addressing request/callback and transaction fixes (EUCA-4172)
AsyncRequests now has a dispatchSafely method that can be used if the
caller wishes to ensure that either the success or failure case of a
requests callback is run.

The Request interface and AsyncRequest implementation are updated to use
more appropriate type information. Address callbacks now use type
parameters so the compiler can validate types.

Addresses now exposes methods for updating instance IP addresses within
transactions.

The duplicated system address assignment logic is moved to
AbstractSystemAddressManager.doAssignSystemAddress where it can be used
by the DynamicSystemAddressManager and StaticSystemAddressManager.
System address assignment is updated to use a transaction when updating
the public address for an instance post assignment.

UnassignAddressCallback is updated to use a transaction when updating
the public address for an instance on success or failure.

Fixes EUCA-4172