Address release on terminate when cluster controller offline (EUCA-2295)
It is now safe to call 'abort' for a ResourceToken multiple times, the
first call will free resources and subsequent calls are ignored. Prior
to this fix multiple calls to abort may have resulted in incorrectly
freeing resources that were no longer allocated to the token.

For some instance state changes (e.g. running to stopping or stopping to
terminated) we attempt to clean up resources for the instance, which
includes clean up of the public IP assignment. A message is sent to the
cluster controller to remove the IP address assignment, but if there is
no enabled cluster controller for the partition then neither the success
or error callbacks were being invoked.

The fix ensures that when an address is unassigned we perform an action
for either the success or error cases. When an instance is terminated
(moves to the shutting_down or stopping states) we will preserve the
public IP address association if the cluster controller cannot be
reached. Should this occur then when the instance later transitions to
the terminated state we will unassign the public IP address.

Fixes EUCA-2295