Add filtering on resource properties for key pairs (EUCA-4674)
Key pair filtering is supported by the new KeyPairFilterSupport class in
KeyPairs. A FilterFunctions enum is added to KeyPairs with a value for
each (String) property. KeyPairs is updated to allow use of a Criteria /
alias Map for key pair listing (for databasefilters). KeyPair manager is
updated to support filtering via the new CloudMetadatas API.
KeyPairFilterSupportTest is added to provide coverage for all
Predicate filters and validation of filtering configuration.