Add support for AWS Signature V4 (EUCA-4648)
QueryPipeline now allows the required request parameters to be
specified and tracks which parameters were from form fields or the query
string. EuareQueryPipeline is updated to require only a Version
parameter as Signature V4 does hot require any query parameters.

HmacCredentials is updated to remove any assumptions about the location
of values such as the access key identifier and signature method
declaration. HmacLoginModuleSupport is updated to allow specialization
of various signature version specific tasks.

The HmacUtils class is moved and expanded to handle signature versions
and their variants (Signature V4 has a variation that uses query
parameters)

HmacHandler is updated to delegate signature version specific tasks to
the signature variant in use.

QueryTimestampHandler is updated to use HmacUtils as a facade hiding the
details of the signature variant in use.

TokensQueryPipeline is updated to account for Signature V4 credentials
when checking whether to use password authentication (for user console
functionality)

A SignatureHandlerUtils class is added to provide lookup of parameter
and header values from a MappingHttpRequest.

Existing HMAC login modules are updated to ignore parameters as
appropriate and to accept multiple values for each parameter when
canonicalizing. An Hmacv4LoginModule is added for login with Signature
V4, this module does not currently validate the region or service and
has a special case to allow for AWS Java SDK signatures with an
incorrect path of '/'.

The HmacLoginModuleTest is updated with tests for Signature V4, this
includes running the AWS provided test kit and a test using a request
generated by the AWS Java SDK version 1.3.26.

A new SecurityHeader enum is added for working with HTTP headers related
to security and the SecurityParameter enum is updated to support the new
parameters for the query variant of Signature V4.

Timestamps is updated to support more timestamp formats. Short variants
of ISO 8601 dates are now accepted and RFC 2616 / HTTP 1.1 can now also
be parsed (e.g. from the HTTP 'Date' header)

MappingHttpRequest is updated to support identification of parameters
that were from form fields (as opposed to the query string), this is
necessary as form field parameters should not be included in the
canonical form of the query parameters (they are included as the request
body)

A CollectionUtils class is added for any utility methods related to
collections. A Strings class is added for utility methods related to
Strings.