Resource tags are now included when describing resources  (EUCA-4749)
The response message to the describe operation for images, instances,
security groups, snapshots and volumes now include the resource tags
for the requesting users account.

The 10-08-31 binding is updated to include tags along with the
associated message data object for each type. Service classes now use
the new 'getResourceTagMap' TagSupport method to bulk load the tags and
the 'addFromTags' Tags utility method to add tags to the response.

Unit tests are added to validate the class details used in Hibernate
criteria queries for loading tags.