Update exception handling framework for web services.
A new EucalyptusWebServiceException class is added for extention by
service specific exceptions. The new ErrorHandlerSupport class uses the
exception information to support handling of service errors.

The new QueryBindingInfo annotation allows error sub-classes to declare
the HTTP status code for use with that error condition. A Role
enumeration is added to identify the party at fault for an error
condition.

BaseMessageSupplier is added to encapsulate a message and any associated
binding details (initially just the HTTP status code) and web services
framework code is updated to allow responses of this type.