Non tag resource filtering support for instances - EUCA-4672
VmInstance is updated to extract display value calculations from the
instance describe type mapper, this allows filters to re-use these
values.

VmInstances is updated to add filters for instances to the
VmInstanceFilterSupport class. Enums are added for the functions related
to the filters.

FilterSupportTest is updated to fix alias support and to allow for
relations using sub-types.

VmInstanceFilterSupportTest is updated to cover predicate filters where
possible to do so without mocking support.

EUCA-4672