Auto scaling group management - EUCA-4756
A "groups" package is added with auto scaling group persistent objects
and an AutoScalingGroups class providing access. Metadata exceptions are
added for auto scaling to remove use of ones from the cloud manager
module.

The AutoScalingService is updated to support create, describe, update
and delete for auto scaling groups and setting of desired capacity.
Describe of termination policy types is also implemented. Associated
messages are updated with convenience methods and Query API binding
annotations. An InvalidParameterValueException is added for use with
incorrect parameters.

Added a QuantityMetricFunction and AutoScalingGroupNumberQuotaKey for
auto scaling group quota support. Common quota key functionality is
moved to the new AutoScalingNumberQuotaKeySupport class.

EUCA-4756