Auto scaling launch and terminate activities - EUCA-4893
Auto scaling is updated with support for launch and termination of
instances. Launch support includes tagging instances with the group name
and termination of instances includes evaluation of the relevant
termination policies. Launch and terminate activities are subject to
back off on failure and a configuration property is added to disable all
scaling activities in the cloud.

The AutoScalingGroup entity is updated to track the current capacity of
each group and if scaling is currently required for the group.

Entities are added for scaling activities and instances and a new
ActivityManager is added to direct/coordinate scaling activities. The
ActivityManager uses the new BackoffRunner for handling of back off on
failures and the new EucalyptusClient for access to the Eucalyptus (EC2)
service.

TerminationPolicyType is updated with implementations for each policy. A
TerminationPolicyTypeTest unit test is added for validation.

A new InternalImpersonationHandler is added and ServiceContext updated
to support impersonation for remote and local clients respectively (this
is subject to change)

TagManager is updated to allow creation of tags with restricted names
for internal clients.

EUCA-4893