Auto scaling terminate auto scaling instance - EUCA-4899
Support is added for termination of an auto scaling instance and for
forcing termination of all instances when deleting an auto scaling
group.

The AutoScalingService uses the ActivityManager for calling out to EC2
and to handle checking for existing scaling activity for the group. The
BackoffRunner will allow user requested termination activities to run
even when all other scaling activities are disabled.
AutoScalingInstances is updated to allow listing and deletion of
instances for an auto scaling group. A ScalingActivityTransform type
mapper is added to create the response for the instance termination
operation.

AutoScalingService errors are updated, the correct error is now returned
when attempting to delete a resource that is in use. There is a new
error for scaling activity in progress.

LifecycleState is updated with the correct values but these are not yet
used.

EUCA-4899