Auto Scaling support for scaling activity cooldown - EUCA-4989
The auto scaling service now checks the cooldown period when setting the
desired capacity or executing a policy when HonorCooldown is true. Auto
scaling groups now record the time the desired capacity was last
updated.

The auto scaling service is updated to fail operations as appropriate if
scaling is in progress. The ActivityManager and BackoffRunner now
support checking if a task / activity is in progress.

EUCA-4989